update zmq socket in send_joint_state.py

+++ src/vla_center/vla_center/send_joint_state.py
@@ -74,7 +74,7 @@ class CurrentJointStateSenderNode(Node):
         buf = io.BytesIO()
         np.savez_compressed(
             buf,
-            ts=np.array([ts], dtype=np.float64),
+            ts=np.array([ts], dtype=np.float32),
             joints=joints,
         )
         return buf.getvalue()
